app/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
pyperclip>=1.8.2
streamlit-drawable-canvas>=0.9.0
opencv-python>=4.8.0

streamlit==1.37.1
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2

streamlit-drawable-canvas==0.9.3
